--- ai/requirements.txt
@@ -1,7 +1,0 @@
-fastapi
-uvicorn[standard] 
-langchain
-langchain-google-genai
-scikit-learn
-numpy
-python-dotenv
